+++ crush/ui/fs_panel.py
@@ -439,10 +439,12 @@ class FilesystemPanel(QWidget):
         open_proto_action = open_as_menu.addAction("Protobuf")
         open_realm_encrypted_action = None
         open_sqlcipher_action = None
+        open_pdf_encrypted_action = None
         if not node.is_dir:
             open_as_menu.addSeparator()
             open_realm_encrypted_action = open_as_menu.addAction("Realm DB (Encrypted)…")
-            open_sqlcipher_action = open_as_menu.addAction("SQLite DB (SQLCipher)…")
+            open_sqlcipher_action = open_as_menu.addAction("SQLite DB (Encrypted)…")
+            open_pdf_encrypted_action = open_as_menu.addAction("PDF (Encrypted)…")
         open_logs_folder_action = None
         open_multi_log_action   = None
         add_multi_log_action    = None
@@ -513,6 +515,8 @@ class FilesystemPanel(QWidget):
             self.open_requested.emit(node, vfs, "realm_encrypted")
         elif action == open_sqlcipher_action:
             self.open_requested.emit(node, vfs, "sqlcipher")
+        elif action == open_pdf_encrypted_action:
+            self.open_requested.emit(node, vfs, "pdf_encrypted")
         elif action == open_external_default:
             self.open_external_requested.emit(node, vfs, "default")
         elif action == open_external_choose: